+++ App/View/Screens/GenericChat.js
@@ -10,18 +10,16 @@ import {
     ActivityIndicator,
     TouchableOpacity,
     RefreshControl,
-    Image,
     Dimensions
 } from 'react-native';
 import firestore from '@react-native-firebase/firestore';
 import auth from '@react-native-firebase/auth';
 import Icon from 'react-native-vector-icons/Ionicons';
 import { SafeAreaView } from 'react-native';
-import Hyperlink from 'react-native-hyperlink'
-import { resolvePreset } from '@babel/core';
-import { AutoGrowingTextInput } from 'react-native-autogrow-textinput';
+import ChatMessage from '../Components/ChatMessageComponent'
 
-let msgToLoad = 15;
+
+let msgToLoad = 20;
 let msgToStart = 0;
 let endReached = false;
 let flag = false;
@@ -31,50 +29,7 @@ const ListFooterComponent = () => {
 };
 //this function dicompose the doc of the coplete chat into small object where as any object represent one msg item inside our chat
 //
-ChatMessage = ({ item }) => {
-    let date = item.date.toDate();
 
-    return (
-        <View
-            style={
-                auth().currentUser && auth().currentUser.email === item.user_id
-                    ? styles.myItemElement
-                    : styles.ItemElement
-            }>
-            <View>
-                <Image
-                    style={styles.userPhoto}
-                    source={{
-                        uri: 'https://cdn4.iconfinder.com/data/icons/ionicons/512/icon-ios7-contact-512.png',
-                    }}></Image>
-            </View>
-            <View
-                style={
-                    auth().currentUser && auth().currentUser.email === item.user_id
-                        ? styles.myMessageBox
-                        : styles.messageBox
-                }>
-                <Hyperlink linkDefault={ true } linkStyle={ { color: '#2980b9'} }>
-                <Text style={styles.messageStyle}>{item.message}</Text>
-                </Hyperlink>
-                <View style={styles.messageDetails}>
-                    <Text style={styles.userId}>{' ' + item.user_nick + ' '}</Text>
-                    <Text style={styles.date}>
-                        {' ' +
-                            date.getDate() +
-                            '/' +
-                            (date.getMonth() + 1) +
-                            ' ' +
-                            (date.getHours() + 3) +
-                            ':' +
-                            ('0' + date.getMinutes()).slice(-2) +
-                            ' '}
-                    </Text>
-                </View>
-            </View>
-        </View>
-    );
-};
 
 GenericChat = ({ navigation, route }) => {
     const [newMessage, setNewMessage] = useState('');
@@ -98,7 +53,6 @@ GenericChat = ({ navigation, route }) => {
     sendMessage = () => {
         let tempMessage = {
             user_id: user.email,
-            user_nick: user.name,
             message: newMessage,
             date: new Date(),
         };
@@ -127,7 +81,7 @@ GenericChat = ({ navigation, route }) => {
     // we read all the chat list from the chat and we slice it into the amount of messages we want to see.
     function loadMore(chat_data) {
         set_loading_more(true);
-        msgToLoad = msgToLoad + 7;
+        msgToLoad = msgToLoad + 20;
         firestore()
             .collection('chats')
             .doc('reporters')
@@ -311,60 +265,7 @@ const styles = StyleSheet.create({
         fontSize: 11,
         color: '#333333',
     },
-    messageBox: {
-        margin: 5,
-        padding: 5,
-        backgroundColor: '#fff',
-        borderColor: 'black',
-        // borderWidth: 1,
-        borderRadius: 15,
-        alignSelf: 'flex-start',
-        maxWidth: '83%',
-    },
-    myMessageBox: {
-        margin: 5,
-        padding: 5,
-        backgroundColor: 'rgb(205,255,230)',
-        borderColor: 'black',
-        // borderWidth: 1,
-        borderRadius: 15,
-        alignSelf: 'flex-start',
-        maxWidth: '83%',
-        shadowColor: "#00f",
-        shadowOffset: {
-            width: 0,
-            height: 5,
-        },
-        shadowOpacity: 0.36,
-        shadowRadius: 5,
-
-        elevation: 5,
-    },
-    ItemElement: {
-        flexDirection: 'row',
-        flex: 1,
-        alignContent: 'flex-end',
-        flexDirection: 'row-reverse',
-    },
-    myItemElement: {
-        flexDirection: 'row',
-        flex: 1,
-        alignContent: 'flex-end',
-    },
-    messageStyle: {
-        color: 'black',
-        fontSize: 17,
-        paddingLeft: 4,
-    },
-    userPhoto: {
-        width: 30,
-        height: 30,
-        margin: 5,
-    },
-    messageDetails: {
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-    },
+    
 });
 
 export default GenericChat;
